move functions to movegen / board
