+++ src/board.h
@@ -57,3 +57,17 @@ INLINE int Distance(const Square sq1, const Square sq2) {
 INLINE int RelativeRank(const Color color, const int rank) {
     return color == WHITE ? rank : RANK_8 - rank;
 }
+
+INLINE bool ValidSquare(const Square sq) {
+    return sq <= H8;
+}
+
+INLINE bool ValidSide(const Color side) {
+    return side == WHITE
+        || side == BLACK;
+}
+
+INLINE bool ValidPiece(const Piece piece) {
+    return (wP <= piece && piece <= wK)
+        || (bP <= piece && piece <= bK);
+}
