--- src/validate.h
@@ -1,29 +1,0 @@
-/*
-  Weiss is a UCI compliant chess engine.
-  Copyright (C) 2020  Terje Kirstihagen
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*/
-
-#pragma once
-
-#include "types.h"
-
-
-#ifndef NDEBUG
-bool ValidSquare(Square sq);
-bool ValidSide(Color side);
-bool ValidPiece(Piece piece);
-bool MoveListOk(const MoveList *list, const Position *pos);
-#endif
